// src/app/model/user.ts
export class User{
    /*
    id: string | undefined;
    userName: string | undefined
    password: string | undefined
    firstName: string | undefined
    lastName: string | undefined
    token: string | undefined
    */
    id: string="";
    userName: string ="";
    password: string ="";
    firstName: string ="";
    lastName: string ="";
    email:string="";
    role:string="";
    token: string ="";

}